+++ config/types/TMDBMovieDetails.ts
@@ -1,0 +1,51 @@
+export interface TMDBMovieDetails {
+    adult: boolean;
+    backdrop_path: string | null;
+    belongs_to_collection: null | Record<string, any>; // Adjust if you know the collection structure
+    budget: number;
+    genres: {
+      id: number;
+      name: string;
+    }[];
+    homepage: string;
+    id: number;
+    imdb_id: string;
+    origin_country: string[];
+    original_language: string;
+    original_title: string;
+    overview: string;
+    popularity: number;
+    poster_path: string | null;
+    production_companies: {
+      id: number;
+      logo_path: string | null;
+      name: string;
+      origin_country: string;
+    }[];
+    production_countries: {
+      iso_3166_1: string;
+      name: string;
+    }[];
+    release_date: string; // ISO 8601 date string
+    revenue: number;
+    runtime: number;
+    spoken_languages: {
+      english_name: string;
+      iso_639_1: string;
+      name: string;
+    }[];
+    status: string;
+    tagline: string;
+    title: string;
+    video: boolean;
+    vote_average: number;
+    vote_count: number;
+    external_ids: {
+      imdb_id: string;
+      wikidata_id: string;
+      facebook_id: string | null;
+      instagram_id: string | null;
+      twitter_id: string | null;
+    };
+  }
+